fix(orm): P0-4 schema-from-attributes, remote_side crash, Decimal precision, UniqueConstraints
- Add from_attributes=True to TenantResponse, DepartmentResponse,
  MarketplaceItemResponse, MarketplaceReviewResponse,
  MarketplaceRatingResponse, CloneResponse (Pydantic serialization fix)
- Fix AgentResponse.llm_config alias vs ORM model_config mismatch
  via model_validator bridge
- Fix remote_side string->column object in TenantModel and
  DepartmentModel (mapper crash at startup)
- Sync MarketplaceChangeLogModel columns to match SQL DDL
  (before_snapshot/after_snapshot JSON instead of old_value/new_value)
- Export PublishChannelModel from models/__init__.py
- Fix Float->Numeric(3,2) for ABTestModel.traffic_split and
  PluginModel.rating (DECIMAL precision)
- Add 11 missing UniqueConstraints matching SQL DDL across
  UserModel, UserRoleModel, RoleModel, RolePermissionModel,
  TagModel, TagBindingModel, ConversationVariableModel,
  MessageFeedbackModel, WorkflowNodeModel, AccountIntegrateModel
- Add missing fields to EvaluationRunResponse (duration_ms, avg_scores,
  total_tokens, total_cost, tenant_id, created_by) and
  EvaluationResultResponse (overall_score, token_count, error_message,
  evaluation_id)
- Fix Decimal->float precision in schemas: ConversationResponse,
  WorkflowExecutionResponse, CrewExecutionResponse, all usage schemas
- Add archived_at/deleted_at to ConversationResponse
- Add metadata field to MessageResponse with meta_info ORM mapping
- Upgrade all response schemas to ConfigDict(from_attributes=True)

+++ backend/app/models/extended.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from datetime import datetime, UTC
 from typing import Optional
 
-from sqlalchemy import Column, String, Integer, Float, Boolean, DateTime, JSON, ForeignKey, Text
+from sqlalchemy import Column, String, Integer, Float, Boolean, DateTime, JSON, ForeignKey, Text, Numeric
 from sqlalchemy.orm import relationship
 
 from app.models.base import Base
@@ -22,7 +22,7 @@ class ABTestModel(Base):
     description = Column(Text)
     version_a_id = Column(String(36), ForeignKey("agent_versions.id"), nullable=False)
     version_b_id = Column(String(36), ForeignKey("agent_versions.id"), nullable=False)
-    traffic_split = Column(Float, default=0.5)
+    traffic_split = Column(Numeric(3, 2), default=0.5)
     metric = Column(String(50), default="satisfaction")
     duration_hours = Column(Integer, default=24)
     min_samples = Column(Integer, default=100)
@@ -60,7 +60,7 @@ class PluginModel(Base):
     dependencies = Column(JSON, default=[])
     permissions = Column(JSON, default=[])
     downloads = Column(Integer, default=0)
-    rating = Column(Float, default=0.0)
+    rating = Column(Numeric(3, 2), default=0.0)
     rating_count = Column(Integer, default=0)
     status = Column(String(20), default="draft")  # draft, published, archived
     tenant_id = Column(String(36), nullable=False, index=True)
